pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "marketmaker"
version = "0.1.0"
dependencies = [
    "python-dotenv", # token
    "discord.py",    # discord bot
    "nltk",          # word generation
    "pyenchant",     # spell checking
    "numpy",         # saving int lists
    "pytz",          # timezone
    "discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus", # discord menus
]
requires-python = ">=3.11"
readme = "README.md"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
bot = "marketmaker.bot:run_bot"

[tool.setuptools]
packages = ["marketmaker"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = ["bot: marks tests as requiring a bot token"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "RET504",
    "D100",
    "D205",
    "D401",
    "D212",
    "N806",
    "PLR2004",
    "S101",
    "INP001",
    "PD008",
]